+++ pkg/security/serializers/serializers_others.go
@@ -13,6 +13,7 @@ import (
 
 	"github.com/DataDog/datadog-agent/pkg/security/events"
 	"github.com/DataDog/datadog-agent/pkg/security/secl/model"
+	"github.com/DataDog/datadog-agent/pkg/security/secl/rules"
 )
 
 // EventSerializer serializes an event to JSON
@@ -35,6 +36,6 @@ func MarshalCustomEvent(event *events.CustomEvent) ([]byte, error) {
 }
 
 // NewEventSerializer creates a new event serializer based on the event type
-func NewEventSerializer(_ *model.Event, _ *_.Rule) *EventSerializer {
+func NewEventSerializer(_ *model.Event, _ *rules.Rule) *EventSerializer {
 	return nil
 }
